Implementing CSV feed delay
git-svn-id: file:///space/source.marketcetera.org/root/trunk@16026 501a4a96-914b-0410-8454-f5988a4ca596

+++ source/modules/marketdata/csv/src/main/java/org/marketcetera/marketdata/csv/CSVFeedCredentials.java
@@ -36,14 +36,14 @@ public final class CSVFeedCredentials
     /**
      * Retrieves an instance of <code>CSVFeedCredentials</code>.
      * 
-     * @param inReplayRate a <code>double</code> value containing the rate at which to replay marketdata
+     * @param inReplayRate a <code>long</code> value containing the rate at which to replay marketdata
      * @param inReplayEvents a <code>boolean</code> value indicating whether to replay events upon completion
      * @param inMarketdataDirectory a <code>String</code> value containing the marketdata files
      * @param inEventTranslatorClassname a <code>String</code> value containing the fully-qualified name of the event translator class
      * @return a <code>CSVFeedCredentials</code> value
      * @throws FeedException if an error occurs while retrieving the credentials object
      */
-    static CSVFeedCredentials getInstance(double inReplayRate,
+    static CSVFeedCredentials getInstance(long inReplayRate,
                                           boolean inReplayEvents,
                                           String inMarketdataDirectory,
                                           String inEventTranslatorClassname)
@@ -74,13 +74,13 @@ public final class CSVFeedCredentials
     /**
      * Retrieves an instance of <code>CSVFeedCredentials</code>.
      * 
-     * @param inReplayRate a <code>double</code> value containing the rate at which to replay marketdata
+     * @param inReplayRate a <code>long</code> value containing the rate at which to replay marketdata
      * @param inReplayEvents a <code>boolean</code> value indicating whether to replay events upon completion
      * @param inMarketdataDirectory a <code>String</code> value containing the marketdata files
      * @return a <code>CSVFeedCredentials</code> value
      * @throws FeedException if an error occurs while retrieving the credentials object
      */
-    static CSVFeedCredentials getInstance(double inReplayRate,
+    static CSVFeedCredentials getInstance(long inReplayRate,
                                           boolean inReplayEvents,
                                           String inMarketdataDirectory,
                                           CSVFeedEventTranslator inEventTranslator)
@@ -118,9 +118,9 @@ public final class CSVFeedCredentials
     /**
      * Get the replayRate value.
      *
-     * @return a <code>double</code> value
+     * @return a <code>long</code> value
      */
-    public double getReplayRate()
+    public long getReplayRate()
     {
         return replayRate;
     }
@@ -145,7 +145,7 @@ public final class CSVFeedCredentials
     /**
      * Creates a new <code>CSVFeedCredentials</code> instance.
      * 
-     * @param inReplayRate a <code>double</code> value containing the rate at which to replay marketdata
+     * @param inReplayRate a <code>long</code> value containing the rate at which to replay marketdata
      * @param inReplayEvents a <code>boolean</code> value indicating whether to replay events upon completion
      * @param inMarketdataDirectory a <code>String</code> value containing the marketdata files
      * @param inEventTranslatorClassname a <code>String</code> value containing the fully-qualified name of the event translator class
@@ -154,7 +154,7 @@ public final class CSVFeedCredentials
      * @throws InstantiationException if the class referred to by the classname cannot be instantiated
      * @throws FeedException if the given delay is invalid 
      */
-	private CSVFeedCredentials(double inReplayRate,
+	private CSVFeedCredentials(long inReplayRate,
 	                           boolean inReplayEvents,
 	                           String inMarketdataDirectory,
 	                           String inEventTranslatorClassname)
@@ -168,13 +168,13 @@ public final class CSVFeedCredentials
     /**
      * Creates a new <code>CSVFeedCredentials</code> instance.
      * 
-     * @param inReplayRate a <code>double</code> value containing the rate at which to replay marketdata
+     * @param inReplayRate a <code>long</code> value containing the rate at which to replay marketdata
      * @param inReplayEvents a <code>boolean</code> value indicating whether to replay events upon completion
      * @param inMarketdataDirectory a <code>String</code> value containing the marketdata files
      * @param inEventTranslatorClassname a <code>String</code> value containing the fully-qualified name of the event translator class
      * @throws FeedException if an error occurs while constructing the credentials object
      */
-    private CSVFeedCredentials(double inReplayRate,
+    private CSVFeedCredentials(long inReplayRate,
                                boolean inReplayEvents,
                                String inMarketdataDirectory,
                                CSVFeedEventTranslator inEventTranslator)
@@ -197,9 +197,9 @@ public final class CSVFeedCredentials
      */
     private final File marketdataDirectory;
     /**
-     * the number of milliseconds to delay between events
+     * number of milliseconds to delay between events
      */
-    private final double replayRate;
+    private final long replayRate;
     /**
      * indicates whether to replay events upon completion
      */
